requirements.txt
torch>=1.9.0
torchvision>=0.10.0
ultralytics==8.0.196
opencv-python>=4.5.0
matplotlib>=3.3.0
numpy>=1.21.0
pillow>=8.3.0
tqdm>=4.62.0
pycocotools
seaborn>=0.11.0
pandas>=1.3.0